// src/entities/Profile/model/types/ProfileSchema.ts
import { Country, Currency } from "shared/lib/types/entitiesTypes";

export interface Profile {
    firstname: string;
    lastname: string;
    age: number;
    currency: Currency;
    country: Country;
    city: string;
    username: string;
    avatar: string;
}

export interface ProfileSchema {
    data?: Profile;
    isLoading: boolean;
    isReadonly: boolean;
    error?: string;
}